requirements.txt
Flask==3.0.0
PyGithub==2.1.1
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
psutil==5.9.6
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
